Move SearchForm components
