--- src/components/GlobalHeader/SearchForm/SearchForm.module.scss
@@ -1,49 +1,0 @@
-@import '@styles/mixin';
-.container {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-
-  .search {
-    @include mobile {
-      display: none;
-    }
-    &__input {
-      width: 25rem;
-      border: none;
-      padding: 0.7rem 2rem;
-      border-radius: 100px;
-      transition: all 0.2s;
-      margin-right: -5.5rem;
-      background: var(--bg-secondary);
-      color: var(--text-primary);
-      &::-webkit-input-placeholder {
-        font-weight: 100;
-      }
-      &:focus {
-        outline: none;
-        background: var(--bg-secondary);
-      }
-    }
-
-    &__button {
-      svg {
-        fill: var(--accent);
-      }
-
-      padding-right: 1.5rem;
-      border: none;
-      background-color: transparent;
-      &:focus {
-        outline: none;
-      }
-      &:active {
-        transform: translateY(2px);
-      }
-    }
-    &__icon {
-      height: 1.125rem;
-      width: 1.125rem;
-    }
-  }
-}
