+++ src/components/PostList/TablePostList/Pagination.tsx
@@ -1,4 +1,3 @@
-import Link from 'next/link';
 import { useRouter } from 'next/router';
 import styles from './TablePostList.module.scss';
