+++ src/components/PostList/TablePostList/TablePostList.module.scss
@@ -48,7 +48,7 @@
 .Pagination {
   margin: 0 auto;
   width: fit-content;
-  a {
+  button {
     cursor: pointer;
     font-size: 1.2rem;
     margin: 0 0.4rem;
@@ -57,13 +57,9 @@
     background: none;
   }
   .page {
-    a {
-      color: var(--text-secondary);
-    }
+    color: var(--text-secondary);
   }
   .activePage {
-    a {
-      color: var(--accent);
-    }
+    color: var(--accent);
   }
 }
