+++ src/components/SearchForm/SearchForm.module.scss
@@ -1,0 +1,34 @@
+@import '@styles/mixin';
+.container {
+  margin: 6rem auto 1rem;
+  width: 40rem;
+  @include mobile{
+    width: 100%;
+  }
+  .search {
+    display: flex;
+    border: 1px solid var(--accent);
+    .searchInput {
+      width: 100%;
+      border: none;
+      padding: 0.7rem 2rem;
+      color: var(--text-primary);
+      &:focus {
+        outline: none;
+        background: var(--bg-secondary);
+      }
+    }
+  }
+  .logo {
+    display: flex;
+    justify-content: center;
+    margin: auto 1rem;
+    width: 1.2rem;
+    height: 1.2rem;
+    svg {
+      fill: var(--accent);
+      width: 100%;
+      height: 100%;
+    }
+  }
+}
